created AIA::Prompt from the prompt processing module
