+++ .semver
@@ -1,6 +1,6 @@
 ---
 :major: 0
 :minor: 3
-:patch: 20
+:patch: 21
 :special: ''
 :metadata: ''
